movable_block: move to libtrx and add TR2 walkables support
Resolves #2758.
Resolves #3784.
Resolves #3828.

- add the ability for movable blocks to move on trapdoors,
  drawbridges, bridges, lifts, and falling blocks
- add the ability for stacks of movable blocks to fall and land
  on trapdoors, drawbridges, bridges, lifts, and falling blocks
- add the ability for stacks of movable blocks to fall when on
  opened trapdoors and drawbridges
- add the ability for movable blocks to travel up and down lifts
- fix various bugs with falling movable blocks
- fix Lara hang climbing up a movable block used as a ladder piece
