+++ data/tr2/ship/cfg/tr2-gm/gameflow.json5
@@ -22,6 +22,7 @@
 
     "demo_delay": 30,
     "secret_track": 43,
+    "enable_killer_pushblocks": true,
 
     "title": {
         "path": "data/title_gm.tr2",
