+++ src/libtrx/include/libtrx/game/objects/traps/movable_block.h
@@ -1,18 +1,20 @@
 #pragma once
 
 #include "../../rooms.h"
+#include "../types.h"
 
-void MovableBlock_Initialise(int16_t item_num);
-void MovableBlock_UpdateRotation(ITEM *item, int16_t rot_y);
+// Block or unblock a block's box overlap index.
 void MovableBlock_UpdateBox(const ITEM *item, bool blocked);
-void MovableBlock_SetPushPull(ITEM *item, bool enable);
-bool MovableBlock_IsPushPull(const ITEM *item);
-void MovableBlock_SetGravityFrames(ITEM *item, uint8_t frames);
-uint16_t MovableBlock_GetGravityFrames(const ITEM *item);
-void MovableBlock_SetLinked(ITEM *item);
-GAME_VECTOR MovableBlock_GetLinked(const ITEM *item);
-void MovableBlock_DropStack(
-    int32_t stack_height, XYZ_32 sector_pos, int16_t room_num);
+
+// Drop a stack of blocks.
+void MovableBlock_DropStack(XYZ_32 drop_pos, int16_t room_num);
+
+// Shift a stack of blocks up or down in the y direction.
+void MovableBlock_ShiftStackY(
+    int32_t stack_height, XYZ_32 old_pos, int32_t new_y, int16_t room_num,
+    bool reposition);
+
+// Shift a stack of blocks in the x or z direction..
 void MovableBlock_SlideStack(
     int32_t stack_height, XYZ_32 old_sector, const ITEM *dest_item,
     bool reposition);
